Move time_graph.rs to rustc/util
