+++ src/librustc_codegen_utils/lib.rs
@@ -44,7 +44,6 @@ pub mod link;
 pub mod codegen_backend;
 pub mod symbol_names;
 pub mod symbol_names_test;
-pub mod time_graph;
 pub mod llvm_target_features;
 
 /// check for the #[rustc_error] annotation, which forces an
